write_results_to_json
